# cmd/vulnerability-db-api-cli/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Vulnerability Database API
  description: API for the Vulnerability Database
  version: "1.0"
servers:
- url: https://vulnerabilitydb.vulcan.com
paths:
...
  /findings/{id}:
    get:
      tags:
      - findings
      summary: Get finding findings
      operationId: findings#Get finding
      parameters:
      - name: id
        in: path
        description: Finding ID
        required: true
        schema:
          type: string
          description: Finding ID
          example: Reprehenderit ut non aut odio dolorum omnis.
        example: Soluta aspernatur aut ullam.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
              example:
                details: TLSv1 is enabled and the server supports at least one cipher
                id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                impact_details: The remote service encrypts traffic using an older
                  version of TLS.
                issue:
                  cwe_id: 216
                  description: |-
                    The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                    number of cryptographic design flaws. Modern implementations of TLS 1.0
                    mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                    designed against these flaws and should be used whenever possible.

                    PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                    2018, except for POS POI terminals (and the SSL/TLS termination
                    points to which they connect) that can be verified as not being
                    susceptible to any known exploits.
                  id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                  labels:
                  - Web,TLS
                  recommendations:
                  - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                    1.0.
                  reference_links:
                  - http://www.example.com
                  summary: TLS Version 1.0 Protocol Detection
                resources:
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                score: 7
                source:
                  component: vulcan-tls
                  id: 8502fa58-6066-4890-8039-6cd936885763
                  instance: 6087cba3-fade-4fee-81ca-7459e87de098
                  name: vulcan
                  options: '{"timeout":60}'
                  time: "2019-06-08 11:16:40"
                status: OPEN
                target:
                  id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
                  identifier: www.example.com
                  tags:
                  - sdrn:adevinta:team:security
    patch:
      tags:
      - findings
      summary: Patch finding findings
      operationId: findings#Patch finding
      parameters:
      - name: id
        in: path
        description: Finding ID
        required: true
        schema:
          type: string
          description: Finding ID
          example: Quasi voluptatibus.
        example: Aut deleniti voluptas quis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchFindingRequestBody'
            example:
              status: OPEN
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
              example:
                details: TLSv1 is enabled and the server supports at least one cipher
                id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                impact_details: The remote service encrypts traffic using an older
                  version of TLS.
                issue:
                  cwe_id: 216
                  description: |-
                    The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                    number of cryptographic design flaws. Modern implementations of TLS 1.0
                    mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                    designed against these flaws and should be used whenever possible.

                    PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                    2018, except for POS POI terminals (and the SSL/TLS termination
                    points to which they connect) that can be verified as not being
                    susceptible to any known exploits.
                  id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                  labels:
                  - Web,TLS
                  recommendations:
                  - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                    1.0.
                  reference_links:
                  - http://www.example.com
                  summary: TLS Version 1.0 Protocol Detection
                resources:
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                - attributes:
                  - Hostname
                  - Port
                  - Protocol
                  - Service
                  name: Network Resources
                  resources:
                  - Hostname: adevinta.com
                    Port: "443"
                    Protocol: tcp
                    Service: www
                score: 7
                source:
                  component: vulcan-tls
                  id: 8502fa58-6066-4890-8039-6cd936885763
                  instance: 6087cba3-fade-4fee-81ca-7459e87de098
                  name: vulcan
                  options: '{"timeout":60}'
                  time: "2019-06-08 11:16:40"
                status: OPEN
                target:
                  id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
                  identifier: www.example.com
                  tags:
                  - sdrn:adevinta:team:security
      security:
      - authorization_header_Authorization: []
  /findings/{id}/events:
    get:
      tags:
      - findings
      summary: List events by finding findings
      operationId: findings#List events by finding
      parameters:
      - name: page
        in: query
        description: Number of page requested
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of page requested
          example: 4899692310612258802
          format: int64
        example: 4521982370138960658
      - name: size
        in: query
        description: Number of results per page
        allowEmptyValue: true
        schema:
          type: integer
          description: Number of results per page
          example: 8546609252914259573
          format: int64
        example: 5781645183870503598
      - name: id
        in: path
        description: Finding ID
        required: true
        schema:
          type: string
          description: Finding ID
          example: Eum amet.
        example: Dolor pariatur voluptas.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCollection'
              example:
              - finding_id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                id: 9cf0b778-08cc-49f9-8d57-fa0a560f7522
                score: 7
                source_id: 490eef40-b689-4bf4-bb38-c6a2d16545d1
                time: "2019-06-08 11:16:40"
              - finding_id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                id: 9cf0b778-08cc-49f9-8d57-fa0a560f7522
                score: 7
                source_id: 490eef40-b689-4bf4-bb38-c6a2d16545d1
                time: "2019-06-08 11:16:40"
              - finding_id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                id: 9cf0b778-08cc-49f9-8d57-fa0a560f7522
                score: 7
                source_id: 490eef40-b689-4bf4-bb38-c6a2d16545d1
                time: "2019-06-08 11:16:40"
              - finding_id: 97db1956-b15c-4b50-ae06-fd1e7be57963
                id: 9cf0b778-08cc-49f9-8d57-fa0a560f7522
                score: 7
                source_id: 490eef40-b689-4bf4-bb38-c6a2d16545d1
                time: "2019-06-08 11:16:40"
  /findings/{id}/mttr:
    get:
      tags:
      - findings
      summary: Get finding MTTR findings
      operationId: findings#Get finding MTTR
      parameters:
      - name: id
        in: path
        description: Finding ID
        required: true
        schema:
          type: string
          description: Finding ID
          example: Perferendis quasi aperiam.
        example: Nemo beatae explicabo doloremque voluptatem quis.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mttr'
              example:
                mttr: 305
...
  /labels:
    get:
      tags:
      - labels
      summary: List labels labels
      operationId: labels#List labels
      parameters:
      - name: tag
        in: query
        description: Tag of associated targets
        allowEmptyValue: true
        schema:
          type: string
          description: Tag of associated targets
          example: Corrupti facere et architecto deleniti.
        example: Dolores aperiam ut.
      - name: status
        in: query
        description: Status of associated findings
        allowEmptyValue: true
        schema:
          type: string
          description: Status of associated findings
          example: Iusto commodi commodi at sapiente animi dolores.
        example: Molestias reprehenderit ea sunt reprehenderit nemo.
      - name: atDate
        in: query
        description: Allows to get the list of associated findings for a specific
          date (YYY-MM-DD). This filter has preference over minDate and maxDate.
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get the list of associated findings for a specific
            date (YYY-MM-DD). This filter has preference over minDate and maxDate.
          example: Atque odio ipsam est perspiciatis.
        example: Adipisci rerum earum dolores ut.
      - name: minDate
        in: query
        description: Minimum date filter for associated findings (YYY-MM-DD). This
          filter only applies when used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter for associated findings (YYY-MM-DD). This
            filter only applies when used in conjunction with status filter.
          example: Natus consequatur voluptas culpa eum.
        example: Eaque eos nihil.
      - name: maxDate
        in: query
        description: Maximum date filter for associated findings (YYY-MM-DD). This
          filter only applies when used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter for associated findings (YYY-MM-DD). This
            filter only applies when used in conjunction with status filter.
          example: Quaerat et placeat quae.
        example: Soluta reprehenderit impedit error odio sunt ut.
      - name: identifiers
        in: query
        description: A list of asset identifiers to filter by for associated targets
        allowEmptyValue: true
        schema:
          type: string
          description: A list of asset identifiers to filter by for associated targets
          example: Minus repellat est ut iure.
        example: Molestiae doloremque.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Labels'
              example:
                labels:
                - Web
                - HTTP
                - TLS
...
components:
  schemas:
...
    Finding:
      type: object
      properties:
        details:
          type: string
          description: Finding details
          example: TLSv1 is enabled and the server supports at least one cipher
        id:
          type: string
          description: Finding ID
          example: 97db1956-b15c-4b50-ae06-fd1e7be57963
        impact_details:
          type: string
          description: Finding impact details
          example: The remote service encrypts traffic using an older version of TLS.
        issue:
          $ref: '#/components/schemas/Issue'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceGroup'
          description: Other resources associated with the finding
          example:
          - attributes:
            - Hostname
            - Port
            - Protocol
            - Service
            name: Network Resources
            resources:
            - Hostname: adevinta.com
              Port: "443"
              Protocol: tcp
              Service: www
          - attributes:
            - Hostname
            - Port
            - Protocol
            - Service
            name: Network Resources
            resources:
            - Hostname: adevinta.com
              Port: "443"
              Protocol: tcp
              Service: www
          - attributes:
            - Hostname
            - Port
            - Protocol
            - Service
            name: Network Resources
            resources:
            - Hostname: adevinta.com
              Port: "443"
              Protocol: tcp
              Service: www
          - attributes:
            - Hostname
            - Port
            - Protocol
            - Service
            name: Network Resources
            resources:
            - Hostname: adevinta.com
              Port: "443"
              Protocol: tcp
              Service: www
        score:
          type: number
          description: Finding severity score
          example: 7
          format: float
        source:
          $ref: '#/components/schemas/Sourcewithnotarget'
        status:
          type: string
          description: Finding status. Can be 'OPEN' or 'FIXED'.
          example: OPEN
        target:
          $ref: '#/components/schemas/Target'
      example:
        details: TLSv1 is enabled and the server supports at least one cipher
        id: 97db1956-b15c-4b50-ae06-fd1e7be57963
        impact_details: The remote service encrypts traffic using an older version
          of TLS.
        issue:
          cwe_id: 216
          description: |-
            The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
            number of cryptographic design flaws. Modern implementations of TLS 1.0
            mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
            designed against these flaws and should be used whenever possible.

            PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
            2018, except for POS POI terminals (and the SSL/TLS termination
            points to which they connect) that can be verified as not being
            susceptible to any known exploits.
          id: 234ec73f-74c8-400f-be73-27dae8e5ff16
          labels:
          - Web,TLS
          recommendations:
          - Enable support for TLS 1.1 and 1.2, and disable support for TLS 1.0.
          reference_links:
          - http://www.example.com
          summary: TLS Version 1.0 Protocol Detection
        resources:
        - attributes:
          - Hostname
          - Port
          - Protocol
          - Service
          name: Network Resources
          resources:
          - Hostname: adevinta.com
            Port: "443"
            Protocol: tcp
            Service: www
        - attributes:
          - Hostname
          - Port
          - Protocol
          - Service
          name: Network Resources
          resources:
          - Hostname: adevinta.com
            Port: "443"
            Protocol: tcp
            Service: www
        score: 7
        source:
          component: vulcan-tls
          id: 8502fa58-6066-4890-8039-6cd936885763
          instance: 6087cba3-fade-4fee-81ca-7459e87de098
          name: vulcan
          options: '{"timeout":60}'
          time: "2019-06-08 11:16:40"
        status: OPEN
        target:
          id: 8d14bf4b-2cb1-43f2-9c23-f7985c63858a
          identifier: www.example.com
          tags:
          - sdrn:adevinta:team:security
...
    Labels:
      type: object
      properties:
        labels:
          type: array
          items:
            type: string
            example: Dignissimos molestias unde rerum.
          description: List of labels
          example:
          - Web
          - HTTP
          - TLS
      example:
        labels:
        - Web
        - HTTP
        - TLS
    Mttr:
      type: object
      properties:
        mttr:
          type: integer
          example: 305
          format: int64
      example:
        mttr: 305
    PatchFindingRequestBody:
      type: object
      properties:
        status:
          type: string
          description: Status
          example: OPEN
      example:
        status: OPEN
...
  securitySchemes:
    authorization_header_Authorization:
      type: apiKey
      description: Value must be 'TAG tag=mytag'
      name: Authorization
      in: header
...